proofessor - Copy/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pydantic
python-dotenv
requests
